requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
stripe==7.8.0
djangorestframework-simplejwt[token_blacklist]==5.3.1
python-dotenv==1.0.0
Pillow==10.2.0
requests==2.31.0
django-allauth==0.63.3
cryptography>=42.0.0
